processor/requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.3.0
colorama==0.4.6
craigslistscraper==1.1.2
h11==0.16.0
html5lib==1.1
idna==3.10
inquirerpy==0.3.4
jellyfish==1.2.1
joblib==1.5.2
loguru==0.7.3
lxml==6.0.2
networkx==3.5
nltk==3.9.2
numpy==2.3.4
outcome==1.3.0.post0
packaging==25.0
pfzy==0.3.4
pika==1.3.2
pip-init==0.3.0
plyer==2.1.0
prompt_toolkit==3.0.51
PySocks==1.7.1
python-dotenv==1.1.1
readabilipy==0.3.0
regex==2025.9.18
requests==2.32.4
segtok==1.5.11
selenium==4.34.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
tabulate==0.9.0
termcolor==3.1.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
validators==0.35.0
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0
yake==0.6.0
